make: Add staticcheck, move to tools directory
Add staticcheck to list of tools in the repo, and make it a lint check.

Dev dependencies should be managed in a submodule to keep them out of
the go.mod of the main project. See also uber-go/zap#914.
