+++ .github/workflows/go.yml
@@ -36,7 +36,7 @@ jobs:
           ${{ runner.os }}-go-
 
     - name: Download Dependencies
-      run: go mod download
+      run: make install
 
     - name: Lint
       if: matrix.latest
