+++ Makefile
@@ -1,11 +1,15 @@
 export GOBIN ?= $(shell pwd)/bin
 
-BENCH_FLAGS ?= -cpuprofile=cpu.pprof -memprofile=mem.pprof -benchmem
 GOLINT = $(GOBIN)/golint
+STATICCHECK = $(GOBIN)/staticcheck
 
-GO_FILES := $(shell \
+BENCH_FLAGS ?= -cpuprofile=cpu.pprof -memprofile=mem.pprof -benchmem
+
+GO_FILES = $(shell \
 	find . '(' -path '*/.*' -o -path './vendor' ')' -prune \
 	-o -name '*.go' -print | cut -b3-)
+
+MODULES = . ./tools
 
 .PHONY: all
 all: build lint test
@@ -14,8 +18,15 @@ all: build lint test
 build:
 	go build ./...
 
+.PHONY: install
+install:
+	$(foreach dir,$(MODULES),( \
+		cd $(dir) && \
+		go mod download) && \
+	) true
+
 .PHONY: lint
-lint: $(GOLINT)
+lint: $(GOLINT) $(STATICCHECK)
 	@rm -rf lint.log
 	@echo "Checking formatting..."
 	@gofmt -d -s $(GO_FILES) 2>&1 | tee lint.log
@@ -23,14 +34,19 @@ lint: $(GOLINT)
 	@go vet ./... 2>&1 | tee -a lint.log
 	@echo "Checking lint..."
 	@$(GOLINT) ./... 2>&1 | tee -a lint.log
+	@echo "Checking staticcheck..."
+	@$(STATICCHECK) ./... 2>&1 | tee -a lint.log
 	@echo "Checking for unresolved FIXMEs..."
 	@git grep -i fixme | grep -v -e Makefile | tee -a lint.log
 	@echo "Checking for license headers..."
 	@./check_license.sh | tee -a lint.log
 	@[ ! -s lint.log ]
 
-$(GOLINT):
-	go install golang.org/x/lint/golint
+$(GOLINT): tools/go.mod
+	cd tools && go install golang.org/x/lint/golint
+
+$(STATICCHECK): tools/go.mod
+	cd tools && go install honnef.co/go/tools/cmd/staticcheck
 
 .PHONY: test
 test:
@@ -45,3 +61,7 @@ cover:
 BENCH ?= .
 bench:
 	go list ./... | xargs -n1 go test -bench=$(BENCH) -run="^$$" $(BENCH_FLAGS)
+
+.PHONY: tidy
+tidy:
+	$(foreach dir,$(MODULES),(cd $(dir) && go mod tidy) &&) true
