+++ go.mod
@@ -2,8 +2,4 @@ module go.uber.org/dig
 
 go 1.13
 
-require (
-	github.com/stretchr/testify v1.4.0
-	golang.org/x/lint v0.0.0-20190930215403-16217165b5de
-	golang.org/x/tools v0.0.0-20191030062658-86caa796c7ab // indirect
-)
+require github.com/stretchr/testify v1.4.0
