+++ tools/go.mod
@@ -1,0 +1,11 @@
+module github.com/uber-go/dig/tools
+
+go 1.16
+
+require (
+	github.com/BurntSushi/toml v0.4.1 // indirect
+	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616
+	golang.org/x/sys v0.0.0-20211124211545-fe61309f8881 // indirect
+	golang.org/x/tools v0.1.8 // indirect
+	honnef.co/go/tools v0.2.2
+)
